\name{Aca2-package}
\alias{Aca2-package}
\alias{Aca2}
\docType{package}
\title{
\packageTitle{Aca2}
}
\description{
\packageDescription{Aca2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Aca2}
\packageIndices{Aca2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Aca2}

Maintainer: \packageMaintainer{Aca2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
